# rest-api-doc-generator/openapi.yaml
openapi: 3.1.0
info:
  title: REST API Documentation
  version: 1.0.0
  description: Auto-generated API documentation using AI
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /users:
    get:
      summary: Get all users
      description: Retrieve a list of all users.
      operationId: getUsers
      tags:
        - Users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: User ID
                    name:
                      type: string
                      description: User Name
                    email:
                      type: string
                      description: User Email
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
    post:
      summary: Create a new user
      description: Creates a new user record.
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's name
                email:
                  type: string
                  description: User's email address
                password:
                  type: string
                  description: User's password
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The newly created user's ID
                  name:
                    type: string
                    description: User's name
                  email:
                    type: string
                    description: User's email address
        '400':
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message describing the validation failure
  /users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve detailed information about a specific user.
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation - user found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The user's ID.
                  name:
                    type: string
                    description: The user's name.
                  email:
                    type: string
                    description: The user's email address.
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message indicating user not found.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message indicating an internal server error.
    put:
      summary: Update user by ID
      description: Update an existing user by their ID. Requires authentication.
      operationId: updateUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to update
          schema:
            type: string
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The ID of the updated user
                  name:
                    type: string
                    description: The updated user's name
                  email:
                    type: string
                    description: The updated user's email
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message describing the invalid input
    delete:
      summary: Delete a user by ID
      description: Delete a user from the system. Requires admin privileges.
      operationId: deleteUserById
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID to delete
          schema:
            type: string
      responses:
        '204':
          description: User successfully deleted. No content returned.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message
  /products:
    get:
      summary: Get all products
      description: Retrieves a list of all products.
      operationId: getProducts
      tags:
        - Products
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Product ID
                    name:
                      type: string
                      description: Product Name
                    price:
                      type: number
                      format: float
                      description: Product Price
                    description:
                      type: string
                      description: Product Description
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
    post:
      summary: Create a new product
      description: Creates a new product in the system. Requires authentication and product validation.
      operationId: createProduct
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the product
                description:
                  type: string
                  description: A detailed description of the product
                price:
                  type: number
                  format: float
                  description: The price of the product
                category:
                  type: string
                  description: The category the product belongs to
              required:
                - name
                - price
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The ID of the newly created product
                  name:
                    type: string
                    description: The name of the product
                  description:
                    type: string
                    description: A detailed description of the product
                  price:
                    type: number
                    format: float
                    description: The price of the product
                  category:
                    type: string
                    description: The category the product belongs to
        '400':
          description: Bad Request - Invalid product data or validation failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message describing the validation failure
  /products/{id}:
    get:
      summary: Get product by ID
      description: Retrieve detailed information about a specific product
      operationId: getProductById
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  price:
                    type: number
                  imageUrl:
                    type: string
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    patch:
      summary: Update a product
      description: Updates an existing product by ID. Requires authentication.
      operationId: updateProduct
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
components:
  schemas: {}
tags:
  - name: Users
    description: Users related endpoints
  - name: Products
    description: Products related endpoints